fixup! [BTA] Initial API and implementation of `KaptConfiguration`
