+++ compiler/build-tools/kotlin-build-tools-api/src/main/kotlin/org/jetbrains/kotlin/buildtools/api/KotlinToolchains.kt
@@ -12,6 +12,7 @@ import org.jetbrains.kotlin.buildtools.api.internal.KotlinCompilerVersion
 import org.jetbrains.kotlin.buildtools.api.internal.wrappers.KotlinWrapperPre2_3_20
 import org.jetbrains.kotlin.buildtools.api.internal.wrappers.KotlinWrapperPre2_4_0
 import org.jetbrains.kotlin.buildtools.api.internal.wrappers.KotlinWrapperPre2_4_20
+import org.jetbrains.kotlin.buildtools.api.internal.wrappers.KotlinWrapperPre2_5_0
 import org.jetbrains.kotlin.buildtools.api.js.JsPlatformToolchain
 import org.jetbrains.kotlin.buildtools.api.jvm.JvmPlatformToolchain
 import org.jetbrains.kotlin.buildtools.api.wasm.WasmPlatformToolchain
@@ -171,6 +172,9 @@ public interface KotlinToolchains {
             }
             if (kotlinCompilerVersion < KotlinCompilerVersion(2, 4, 20, "snapshot")) {
                 baseImplementation = KotlinWrapperPre2_4_20(baseImplementation)
+            }
+            if (kotlinCompilerVersion < KotlinCompilerVersion(2, 5, 0, "snapshot")) {
+                baseImplementation = KotlinWrapperPre2_5_0(baseImplementation)
             }
 
             baseImplementation
