+++ compiler/build-tools/kotlin-build-tools-api/src/main/kotlin/org/jetbrains/kotlin/buildtools/api/internal/wrappers/KotlinWrapperPre2_5_0.kt
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2010-2026 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
+ */
+@file:OptIn(ExperimentalBuildToolsApi::class, DelicateBuildToolsApi::class) @file:Suppress("DEPRECATION")
+
+package org.jetbrains.kotlin.buildtools.api.internal.wrappers
+
+import org.jetbrains.kotlin.buildtools.api.DelicateBuildToolsApi
+import org.jetbrains.kotlin.buildtools.api.ExperimentalBuildToolsApi
+import org.jetbrains.kotlin.buildtools.api.KotlinToolchains
+import org.jetbrains.kotlin.buildtools.api.jvm.JvmPlatformToolchain
+import org.jetbrains.kotlin.buildtools.api.jvm.KaptConfiguration
+import java.nio.file.Path
+
+/**
+ * A wrapper class for `KotlinToolchains` to accommodate functionality
+ * changes and compatibility adjustments for versions pre Kotlin 2.5.0.
+ *
+ * Delegates the majority of functionality to the `base` implementation,
+ * while selectively overriding methods to either introduce new behavior
+ * or adapt existing operations.
+ *
+ * @param base The base implementation of `KotlinToolchains` to wrap.
+ */
+@Suppress("ClassName")
+internal class KotlinWrapperPre2_5_0(
+    val base: KotlinToolchains,
+) : KotlinToolchains by base {
+
+    @Suppress("UNCHECKED_CAST")
+    override fun <T : KotlinToolchains.Toolchain> getToolchain(type: Class<T>): T = when (type) {
+        JvmPlatformToolchain::class.java -> JvmPlatformToolchainWrapper(base.getToolchain(type))
+        else -> base.getToolchain(type)
+    } as T
+
+    private class JvmPlatformToolchainWrapper(private val base: JvmPlatformToolchain) : JvmPlatformToolchain by base {
+        override fun kaptCompilerPluginBuilder(
+            kaptClasspath: List<Path>,
+            stubsOutputDir: Path,
+            sourcesOutputDir: Path,
+            annotationProcessorsClasspath: List<Path>
+        ): KaptConfiguration.Builder {
+            TODO("Only available starting from Kotlin compiler 2.5.0")
+        }
+    }
+}
