+++ compiler/build-tools/kotlin-build-tools-api/src/main/kotlin/org/jetbrains/kotlin/buildtools/api/jvm/JvmPlatformToolchain.kt
@@ -69,6 +69,22 @@ public interface JvmPlatformToolchain : KotlinToolchains.Toolchain {
      */
     public fun discoverScriptExtensionsOperationBuilder(classpath: List<Path>): DiscoverScriptExtensionsOperation.Builder
 
+    /**
+     * Creates a builder for configuring the KAPT (Kotlin Annotation Processing Tool) compiler plugin.
+     *
+     * @param kaptClasspath a list of file paths representing the classpath required for the KAPT compiler plugin.
+     * @param stubsOutputDir a directory where stubs generated by KAPT will be stored.
+     * @param sourcesOutputDir a directory where source files generated by KAPT will be stored.
+     * @param annotationProcessorsClasspath a list of file paths representing the classpath required for annotation processors.
+     * @return a [KaptConfiguration.Builder] instance to further configure the KAPT compiler plugin settings.
+     */
+    public fun kaptCompilerPluginBuilder(
+        kaptClasspath: List<Path>,
+        stubsOutputDir: Path,
+        sourcesOutputDir: Path,
+        annotationProcessorsClasspath: List<Path>,
+    ): KaptConfiguration.Builder
+
     public companion object {
         /**
          * Gets a [JvmPlatformToolchain] instance from [KotlinToolchains].
