+++ compiler/build-tools/kotlin-build-tools-api/src/main/kotlin/org/jetbrains/kotlin/buildtools/api/jvm/operations/JvmCompilationOperation.kt
@@ -11,7 +11,6 @@ import org.jetbrains.kotlin.buildtools.api.internal.BaseOption
 import org.jetbrains.kotlin.buildtools.api.jvm.JvmIncrementalCompilationConfiguration
 import org.jetbrains.kotlin.buildtools.api.jvm.JvmPlatformToolchain
 import org.jetbrains.kotlin.buildtools.api.jvm.JvmSnapshotBasedIncrementalCompilationConfiguration
-import org.jetbrains.kotlin.buildtools.api.jvm.KaptConfiguration
 import org.jetbrains.kotlin.buildtools.api.trackers.CompilerLookupTracker
 import java.nio.file.Path
 import kotlin.contracts.ExperimentalContracts
@@ -146,22 +145,6 @@ public interface JvmCompilationOperation : BaseCompilationOperation, Cancellable
             dependenciesSnapshotFiles: List<Path>,
             shrunkClasspathSnapshot: Path,
         ): JvmSnapshotBasedIncrementalCompilationConfiguration.Builder
-
-        /**
-         * Creates a builder for configuring the KAPT (Kotlin Annotation Processing Tool) compiler plugin.
-         *
-         * @param kaptClasspath a list of file paths representing the classpath required for the KAPT compiler plugin.
-         * @param stubsOutputDir a directory where stubs generated by KAPT will be stored.
-         * @param sourcesOutputDir a directory where source files generated by KAPT will be stored.
-         * @param annotationProcessorsClasspath a list of file paths representing the classpath required for annotation processors.
-         * @return a [KaptConfiguration.Builder] instance to further configure the KAPT compiler plugin settings.
-         */
-        public fun kaptCompilerPluginBuilder(
-            kaptClasspath: List<Path>,
-            stubsOutputDir: Path,
-            sourcesOutputDir: Path,
-            annotationProcessorsClasspath: List<Path>,
-        ): KaptConfiguration.Builder
     }
 
     /**
